==> Python/1-1.City_Geo.ipynb <==
 # Dependencies
import pandas as pd
import numpy as np
import requests
import json

# Google API Key
from config import gkey

 # Import cities file as DataFrame
cities_pd = pd.read_csv("cities.csv")
cities_pd

# Add columns for lat, lng, airport name, airport address, airport rating
# Note that we used "" to specify initial entry.
cities_pd["Lat"] = ""
cities_pd["Lng"] = ""
cities_pd["Zip Code"] = ""
cities_pd["Population"] = ""
cities_pd.head()

 # create a params dict that will be updated with new city each iteration
params = {"key": gkey}

# Loop through the cities_pd and run a lat/long search for each city
for index, row in cities_pd.iterrows():
    base_url = "https://maps.googleapis.com/maps/api/geocode/json"

    city = row['City']
    

    # update address key value
    params['address'] = f"{city}"

    # make request, print url
    cities_lat_lng = requests.get(base_url, params=params)
    print(cities_lat_lng.url)
    # convert to json
    cities_lat_lng = cities_lat_lng.json()

    cities_pd.set_value(
        index, "Lat", cities_lat_lng["results"][0]["geometry"]["location"]["lat"])
    cities_pd.set_value(
        index, "Lng", cities_lat_lng["results"][0]["geometry"]["location"]["lng"])

# Visualize to confirm lat lng appear
cities_pd.head()

# Save Data to csv
cities_pd.to_csv("CitiesGeo_Output.csv")

# Visualize to confirm airport data appears
cities_pd.head(10)
